requirements.txt

google-generativeai==0.8.3
python-dotenv==1.0.0
requests==2.31.0
PyGithub==2.1.1